Added statistics command.
