--- shop.js
@@ -1,2 +1,0 @@
-//This feature, coming soon, will be a shop where customized items, tags, etc. can be "bought" using the bot's Sif Coins currency.
-//The development of this shop will be heavily dependent on another project I am working on called Webset, which will be the hosted API making this possible
